feat: move common package out from the lobby server folder
