+++ bot_manager/redis_store.go
